+++ PF-Henry/client/src/Components/Home/Home.jsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import DropdownsFiltros from "../DropdownsFiltros/DropdownsFiltros.jsx";
 import Footer from "../Footer/footer.jsx";
 import NavBar from "../NavBar/NavBar";
-import Card from "../Card/Card";
+import Card_Grid from "../Card_Grid/Card_Grid";
 import Carousel from "../Carousel/Carousel";
 
 //import ShoppingCart from "../ShoppingCart/ShoppingCart.jsx";
@@ -18,7 +18,7 @@ export default function Home() {
             <Carousel />
 
             <div className="container  tw-rounded-lg">
-                <Card />
+                <Card_Grid />
             </div>
         </div>
     );
